requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
httpx==0.27.0